requirements.txt
fastapi
uvicorn
selenium
undetected-chromedriver
beautifulsoup4
requests
pydantic
setuptools
webdriver-manager
lxml
python-dotenv
